feat: add dotenv python package && .flaskenv file
Added the dotenv python package so that Flask always knows that microblog.py is that main
application that needs to be run when starting the web server.  In this case,
the environment variable (FLASKAPP=) is assigned to microblog.py in the .flaskenv file.
Flask will look at this environment variable in the .flaskenv file to know which application
it needs to start.