feat: import render_template and render index.html
Imported the render_template() function from the flask module. This
function will accept the name of a webpage from the templates folder
as an argument and Flask will render it when the app is run.

Added render_template('index.html') to the index view to render
index.html when the app is run.